refactor: create type file
