+++ src/config/keys.ts
@@ -10,7 +10,7 @@ import {
 } from '@graasp/sdk';
 import { DEFAULT_LANG } from '@graasp/translations';
 
-import { ItemSearchParams } from '../item/routes.js';
+import { ItemSearchParams } from '../item/types.js';
 import { PaginationParams } from '../types.js';
 import { AggregateActionsArgs } from '../utils/action.js';
 import { DEFAULT_THUMBNAIL_SIZE } from './constants.js';
