+++ src/item/routes.ts
@@ -1,41 +1,15 @@
-import {
-  ItemType,
-  Member,
-  PermissionLevel,
-  UUID,
-  UnionOfConst,
-} from '@graasp/sdk';
+import { UUID } from '@graasp/sdk';
 
 import qs from 'qs';
 
 import { DEFAULT_THUMBNAIL_SIZE } from '../config/constants.js';
 import { PaginationParams } from '../types.js';
+import { ItemChildrenParams, ItemSearchParams } from './types.js';
 
 export const ITEMS_ROUTE = 'items';
 export const GET_OWN_ITEMS_ROUTE = `${ITEMS_ROUTE}/own`;
 export const GET_RECYCLED_ITEMS_DATA_ROUTE = `${ITEMS_ROUTE}/recycled`;
 export const SHARED_ITEM_WITH_ROUTE = `${ITEMS_ROUTE}/shared-with`;
-
-export type ItemSearchParams =
-  | {
-      creatorId?: Member['id'];
-      name?: string;
-      ordering?: 'desc' | 'asc';
-      sortBy?:
-        | 'item.name'
-        | 'item.type'
-        | 'item.creator.name'
-        | 'item.created_at'
-        | 'item.updated_at';
-      permissions?: PermissionLevel[];
-      types?: UnionOfConst<typeof ItemType>[];
-    }
-  | undefined;
-
-export type ItemChildrenParams = {
-  ordered?: boolean;
-  types?: UnionOfConst<typeof ItemType>[];
-};
 
 export const buildGetAccessibleItems = (
   params: ItemSearchParams,
